2017 Day 3 Part 2
Also fixes the core formula for part one, which was right about half the time (and this included my input value), but was wrong all the rest of the time and part two needed it to be right all the time. So it now has some tests, which would have spotted this problem earlier.

Beware of operator precedence! Unary - binds less tightly than method calls.