2019 Make Intcode computer generic over the value type
Some serious constraints on it as we have to use these values to index into Vec, so it has to be convertible to usize, and it has to be able to participate in various mathematical operations. Should work with any signed integer primitive, and anything else implementing the required traits in a non-surprising manner.

No other code has been updated to use it yet, so this commit renders the solution unable to compile. Existing days to be modified to request Program<i32> which will match their previous behaviour.